recovery acls

+++ rkit.pl
@@ -33,29 +33,35 @@ my $cfg = new Config::Simple("$confdir\\init.conf") or die "File $confdir\\init.
 my $url = $cfg->param('url');
 my $pass = $cfg->param('pass');
 
-my $bkitDir = '.bkit.pt';
+my $bkitDir = '.bkit';
 my $bkit = "$drive:\\$bkitDir";
 -d $bkit or mkdir $bkit;
 my ($logs,$perms,$vols) = map {-d $_ or mkdir $_; $_} map {"$bkit\\$_"} qw(logs perms vols);
-my $acls = "$perms\\acls.txt";
 
 my $fmt = q#'"%p|%t|%o|%i|%b|%l|%f"'#;
-if (defined $drive && defined $backup && defined $computer){
+if (defined $drive && defined $backup && defined $computer && defined $entry){
 	my $lpath = "$drive:/$path";
+	my $logfile = "${logs}\\recv.log";
+	my $acls = "$perms/acls.txt";
+	$acls =~ s#\\+#/#g;															#dos2linux
+	$acls =~ s#^[a-z]:/?##gi;															#dos2linux
+
 	eval{
 		-d $lpath or make_path $lpath; 
-		$lpath =~ s#/#\\#g;												          #linux2dos
+		$lpath =~ s#/+#\\#g;												    #linux2dos
 		my $push = "$perl $cd\\bkit.pl $lpath\\$entry";							#First backup it to server
 		print qx|${push} 2>&1|;
-		$? == 0 or die "The command $push exit with non zero value:$?\nSee file $prerr for details";
+		$? == 0 or die "The command $push exit with non zero value:$?\nSee file ${logfile} for details";
 		my $r = qq|${rsync} -rlitzvvhR --no-perms --delete-delay --delay-updates --force --stats --fuzzy|
 			.qq| --out-format=${fmt}|
-			.qq| ${url}/${drive}/${backup}/./${path}/${entry} /cygdrive/${drive}/|
-			.qq| 1>${logs}\\recv.log 2>&1|;
+			.qq| ${url}/${drive}/${backup}/./${acls}|							#SRC1: data
+			.qq| ${url}/${drive}/${backup}/./${path}/${entry}|					#SRC2: acls
+			.qq| /cygdrive/${drive}/|											#DST
+			.qq| 1>${logfile} 2>&1|;
 		open my $handler, "|-", $r; 									#Now we can restore it
 		print $handler "${pass}\n\n";  
 		print qx|${push} 2>&1|;								#push another backup to server	
-		$? == 0 or die "The command $push exit with non zero value:$?\nSee file $poserr for details";
+		$? == 0 or die "The command $push exit with non zero value:$?\nSee file ${logfile} for details";
 	} or die "Die while executing rsync: $@";
 }
